fix(release): verify every packed source map
