+++ package.json
@@ -29,7 +29,7 @@
     "verify:ui-components-barrel": "node scripts/check-ui-components-barrel.mjs",
     "verify:ui-orientation-selectors": "node scripts/check-ui-orientation-selectors.mjs",
     "verify:date-picker-single-source": "node scripts/check-date-picker-single-source.mjs",
-    "verify:publish-tarballs": "node --test scripts/tests/packed-exports.test.mjs scripts/tests/packed-sourcemaps.test.mjs && node scripts/verify-publish-tarballs.mjs",
+    "verify:publish-tarballs": "node --test scripts/tests/packed-exports.test.mjs scripts/tests/packed-sourcemaps.test.mjs && node scripts/verify-package-tarballs.mjs",
     "verify:tenant-scoping": "node scripts/check-tenant-scoping.mjs",
     "verify:route-authoring": "node scripts/check-route-authoring.mjs",
     "verify:route-ownership": "node scripts/check-route-ownership.mjs",
