+++ packages/runtime-core/tsconfig.build.json
@@ -2,7 +2,7 @@
   "extends": ["./tsconfig.json", "../typescript-config/dep-paths.json"],
   "compilerOptions": {
     "declaration": true,
-    "declarationMap": true,
+    "declarationMap": false,
     "noEmit": false,
     "tsBuildInfoFile": "dist/.tsbuildinfo"
   },
