+++ packages/runtime/tsconfig.build.json
@@ -4,7 +4,7 @@
     "rootDir": "src",
     "outDir": "dist",
     "declaration": true,
-    "declarationMap": true,
+    "declarationMap": false,
     "noEmit": false,
     "tsBuildInfoFile": "dist/.tsbuildinfo"
   },
